Cu fix post frot ènd

+++ ohana-core/src/main/java/vn/ohana/utility/UtilityRepository.java
@@ -9,8 +9,8 @@ import java.util.List;
 import java.util.Set;
 
 @Repository
-public interface UtilityRepository extends JpaRepository<Utility, Long> {
-    List<Utility> findAllByIdIn(Set<Long> ids);
+public interface UtilityRepository extends JpaRepository<Utility, Integer> {
+    List<Utility> findAllByIdIn(Set<Integer> ids);
 
     List<Utility> findAllByStatus(StatusUtility status);
